Fix box load on uninitialized box load
Closes #2376
probably will be rewritten at a later time; am reworking the slot info
messaging for drag/drop/display

+++ PKHeX.Core/Saves/SAV4.cs
@@ -579,7 +579,7 @@ namespace PKHeX.Core
         // Storage
         public override int CurrentBox
         {
-            get => Data[Version == GameVersion.HGSS ? GetBoxOffset(BoxCount) : Box - 4];
+            get => (sbyte)Data[Version == GameVersion.HGSS ? GetBoxOffset(BoxCount) : Box - 4];
             set => Data[Version == GameVersion.HGSS ? GetBoxOffset(BoxCount) : Box - 4] = (byte)value;
         }
